Big chunk of changes necessitated by systemd-timesyncd.service
Because systemd-timesyncd.service has supplanted ntpd in the default
Raspbian image, we need to manually create an ntp user in advance of
testing it.

We also should manually created the /var/lib/ntp directory -
it's easy to overlook the comment within the ntp.conf file to pre-create
this directory. Without the ntp user and /var/lib/ntp directory, the user
will be stuck - ntpq -p won't ever change.

We also should create /var/log/ntpstats, as ntpd throws an error
in the logs regarding the recently discovered default of '/var/NTP' not
existing. At least until that default is fixed. Having that in place
without generating stats is harmless.